modified payment request payload gateway

+++ account/view/subscription_plan_view.py
@@ -1,3 +1,4 @@
+import json
 from typing import cast
 from django.forms import ValidationError
 from django.http import HttpRequest
@@ -106,21 +107,19 @@ class SubscriptionPlanViewSet(viewsets.ModelViewSet):
         request_serializer = RequestSubscriptionSerializer(data=request.data)
         request_serializer.is_valid(raise_exception=True)
         validated_data = request_serializer.validated_data
-
         user = cast(User, request.user)
-        
+
         try:
             with transaction.atomic(): 
                 agent = Agent.get_agent_by_phone_number(phone_number=user.phone_number)
                 plan = SubscriptionPlan.get_plan_by_id(subscription_plan_id=validated_data.get("plan_id"))
 
-                if agent is None:
+                if not agent:
                     return Response(data={"detail": "You are forbidden to view this information"}, status=status.HTTP_404_NOT_FOUND)
 
-                if plan is None:
+                if not plan:
                     return Response(data={"detail": "Subscription plan not found"}, status=status.HTTP_404_NOT_FOUND)
 
-                # Create a payment entry in the database
                 payment = Payment.create(
                     phone_number=validated_data.get("phone_number"),
                     payment_type=PaymentType.ACCOUNT,
@@ -129,59 +128,57 @@ class SubscriptionPlanViewSet(viewsets.ModelViewSet):
                     plan=plan
                 )
 
-                
-                phone_number: str = payment.phone_number
-                
+                phone_number = payment.phone_number
                 if phone_number.startswith('+255'):
                     phone_number = '0' + phone_number[4:]
-                    
+
                 order_data = {
-                    'create_order': str(payment.payment_id),
                     'buyer_email': payment.agent.email,
                     'buyer_name': f"{payment.agent.first_name} {payment.agent.middle_name} {payment.agent.last_name}",
                     'buyer_phone': phone_number,
                     'amount': int(payment.amount),
                     'account_id': settings.ACCOUNT_ID,
-                    'api_key': settings.API_KEY,
-                    'secret_key': settings.KEY,
+                    'webhook_url': f"{settings.APP_BASE}{settings.WEB_HOOK_URL}",
+                    'metadata': json.dumps({
+                        'payment_id': str(payment.payment_id),
+                    }),
+                    'api_key': settings.ZENOPAY_API_KEY,
+                    'secret_key': settings.ZENOPAY_SECRET_KEY,
                 }
 
-                try:
-                    client = HttpClient(base_url=settings.EXTERNAL_API_BASE_URL)
-                    response = client.make_payment(data=order_data)
-                    
-                    logger.info(f"Plan subscription for {payment.agent} with response: {response.text}")
+                client = HttpClient(base_url=settings.ZENOPAY_BASE)
+                response = client.make_payment(data=order_data)
 
-                    if response.status_code == 200:
-                        response_data = response.json()
-                        response_status: str = response_data.get("status")
-                        
-                        if response_status.lower() == "error":
-                            payment.delete()
-                            return Response(data={"detail": response_data.get("message")}, status=status.HTTP_400_BAD_REQUEST)
-                        
-                        payment.update_order_and_message(order_id=response_data.get("order_id"), message=response_data.get("message"))
-                        return Response(data={"detail": f"Successfully subscribed to {plan.name}, please finish your payment to activate account"}, status=status.HTTP_200_OK)
-                except requests.RequestException as e:
-                    logger.error(f"HTTP error occurred: {e}", exc_info=True)
-                    payment.delete()
-                    return Response({"detail": str(e)}, status=status.HTTP_502_BAD_GATEWAY)
-                except requests.Timeout as e:
-                    logger.error(f"Request timeout occurred: {e}", exc_info=True)
-                    payment.delete()
-                    return Response({"detail": str(e)}, status=status.HTTP_504_GATEWAY_TIMEOUT)
-                except Exception as e:
-                    logger.error(f"Unexpected error occurred: {e}", exc_info=True)
-                    payment.delete()
-                    return Response({"detail": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+                if response is None:
+                    return Response({"detail": "Payment gateway error, please try again later"}, status=status.HTTP_502_BAD_GATEWAY)
 
-        
-        except ValueError as e:
-            logger.error(f"Value error occurred: {e}", exc_info=True)
-            return Response({"detail": str(e)}, status=status.HTTP_400_BAD_REQUEST)
-        except ValidationError as e:
-            logger.error(f"Validation error occurred: {e}", exc_info=True)
-            return Response({"detail": str(e)}, status=status.HTTP_400_BAD_REQUEST)
+                logger.info(f"Plan subscription for {payment.agent} with payload: {json.dumps(order_data)} and response: {response.text}")
+
+                if response.status_code == 200:
+                    response_data = response.json()
+                    response_status: str = response_data.get("status")
+
+                    if response_status.lower() == "error":
+                        payment.delete()
+                        return Response(data={"detail": response_data.get("message")}, status=status.HTTP_400_BAD_REQUEST)
+
+                    payment.update_order_and_message(order_id=response_data.get("order_id"), message=response_data.get("message"))
+                    return Response(data={"detail": f"Successfully subscribed to {plan.name}, please complete your payment to activate your account."}, status=status.HTTP_200_OK)
+
+                payment.delete()
+                return Response(data={"detail": "Unexpected error with payment gateway"}, status=status.HTTP_502_BAD_GATEWAY)
+
+        except requests.Timeout as e:
+            logger.error(f"Request timeout occurred: {e}", exc_info=True)
+            payment.delete()
+            return Response({"detail": "Payment request timed out. Please try again later."}, status=status.HTTP_504_GATEWAY_TIMEOUT)
+
+        except requests.RequestException as e:
+            logger.error(f"HTTP error occurred: {e}", exc_info=True)
+            payment.delete()
+            return Response({"detail": "Payment gateway error occurred. Please try again later."}, status=status.HTTP_502_BAD_GATEWAY)
+
         except Exception as e:
             logger.error(f"Unexpected error occurred: {e}", exc_info=True)
+            payment.delete()
             return Response({"detail": f"An unexpected error occurred: {str(e)}"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
